remove SidebarLayout
