--- app/components/layout/SidebarLayout.scss
@@ -1,52 +1,0 @@
-$topBarHeight: 64px;
-
-.component {
-  background: var(--theme-main-body-background-color);
-  height: 100%;
-  display: flex;
-}
-
-.main {
-  width: 100%;
-  height: 100%;
-  overflow: hidden;
-  display: flex;
-  flex-direction: column;
-  position: relative;
-}
-
-.sidebar {
-  flex: 0 0;
-  overflow: visible;
-}
-
-.topbar {
-  height: $topBarHeight;
-  flex-shrink: 0;
-  // Show elements that overflow the topbar above the content area
-  z-index: 1;
-}
-
-.content {
-  height: 100%;
-}
-
-.contentWrapper {
-  height: 100%;
-  overflow-x: hidden;
-  overflow-y: overlay;
-  position: relative;
-  &::-webkit-scrollbar-button {
-    height: 7px;
-    display: block;
-  }
-}
-
-.testnetWarning {
-  font-family: var(--font-bold);
-  text-align: center;
-  color: white;
-  background-color: #b54b4b;
-  padding: 4px;
-  text-transform: uppercase;
-}
