+++ app/components/layout/TopBarLayout.js
@@ -8,6 +8,7 @@ type Props = {
   topbar: Node,
   children?: ?Node,
   notification?: ?Node,
+  banner?: Node
 };
 
 /** Adds a top bar above the wrapped node */
@@ -15,16 +16,20 @@ type Props = {
 export default class TopBarLayout extends Component<Props> {
   static defaultProps = {
     children: undefined,
-    notification: undefined
+    notification: undefined,
+    banner: undefined
   };
 
   render() {
-    const { children, topbar, notification } = this.props;
+    const { banner, children, topbar, notification } = this.props;
     return (
       <div className={styles.component}>
         <div className={styles.topbar}>
           {topbar}
         </div>
+
+        {banner}
+
         {notification}
         <div className={styles.content}>
           {children}
