+++ app/components/topbar/StaticTopbarTitle.scss
