+++ app/components/topbar/WalletTopbarTitle.js
