+++ app/components/topbar/WalletTopbarTitle.scss
