+++ app/components/topbar/banners/TestnetWarningBanner.scss
@@ -1,0 +1,9 @@
+.testnetWarning {
+    font-family: var(--font-bold);
+    text-align: center;
+    color: white;
+    background-color: #b54b4b;
+    padding: 4px;
+    text-transform: uppercase;
+  }
+
