+++ app/containers/daedalusTransfer/DaedalusTransferPage.js
@@ -4,8 +4,8 @@ import { observer } from 'mobx-react';
 import { intlShape, defineMessages } from 'react-intl';
 import validWords from 'bip39/wordlists/english.json';
 import type { InjectedProps } from '../../types/injectedPropsType';
-import StaticTopbarTitle from '../../components/layout/StaticTopbarTitle';
-import TopBar from '../../components/layout/TopBar';
+import StaticTopbarTitle from '../../components/topbar/StaticTopbarTitle';
+import TopBar from '../../components/topbar/TopBar';
 import DaedalusTransferInstructionsPage from '../../components/daedalusTransfer/DaedalusTransferInstructionsPage';
 import DaedalusTransferFormPage from '../../components/daedalusTransfer/DaedalusTransferFormPage';
 import DaedalusTransferWaitingPage from '../../components/daedalusTransfer/DaedalusTransferWaitingPage';
